Stuffs.

+++ Arduino/Main/Main.ino
@@ -23,7 +23,7 @@ ros::Publisher getAngleVel_pub("getAngleVel", &angleVel_msg);
 ros::Subscriber<std_msgs::Float64MultiArray> setTorques_sub("setTorques", &setTorque_callback);
 
 
-static float motorOffsets[5] = {0, M_PI/4.0, M_PI, M_PI * (3.0/4.0), M_PI};
+static float motorOffsets[5] = {0.0, M_PI/4.0, M_PI, M_PI * (3.0/4.0), M_PI};
 
 void setup()
 {
@@ -64,7 +64,7 @@ void setTorque_callback(const std_msgs::Float64MultiArray& msg)
 {
   lastMessageTime = millis();
   
-  long thetas[5];
+  float thetas[5];
   float velocities[5];
 
   for (int i = 1; i < 6; i++)
